Lol floats.

+++ src/main/java/com/dowlandaiello/gitchain/config/ChainConfig.java
@@ -31,7 +31,7 @@ public class ChainConfig {
     public int BlockTime;
 
     /* Starting chain difficulty */
-    public long Difficulty;
+    public Float Difficulty;
 
     /**
      * Initialize a new chain config.
@@ -40,7 +40,7 @@ public class ChainConfig {
      * @param network chain network identifier
      * @param chain chain name / version
      */
-    public ChainConfig(Map<BigInteger, Float> alloc, int network, String chain, int blockTime, long difficulty) {
+    public ChainConfig(Map<BigInteger, Float> alloc, int network, String chain, int blockTime, Float difficulty) {
         this.Alloc = alloc; // Set alloc
         this.Network = network; // Set network
         this.Chain = chain; // Set chain
